refactor: 로직 ui 분리
